Move the metadata
