+++ calyx-frontend/src/metadata.rs
@@ -1,0 +1,123 @@
+use std::{cell::RefCell, collections::HashMap, io::Read, path::PathBuf};
+
+/// An identifier representing a given file path
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct FileId(u32);
+
+/// An identifier representing a location in the Calyx source code
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub struct PositionId(u32);
+
+/// A newtype wrapping a line number
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct LineNum(u32);
+
+impl LineNum {
+    pub fn as_usize(&self) -> usize {
+        self.0 as usize
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct MetadataTable {
+    /// map file ids to the file path, note that this does not contain file content
+    file_map: HashMap<FileId, PathBuf>,
+    /// maps position ids to their source locations. Positions must be handed
+    /// out in order
+    position_map: HashMap<PositionId, SourceLocation>,
+}
+
+impl MetadataTable {
+    pub fn lookup_file_path(&self, file: FileId) -> &PathBuf {
+        &self.file_map[&file]
+    }
+
+    pub fn lookup_position(&self, pos: PositionId) -> &SourceLocation {
+        &self.position_map[&pos]
+    }
+
+    pub fn file_reader(&self) -> MetadataFileReader<'_> {
+        MetadataFileReader::new(self)
+    }
+
+    pub fn add_file(&mut self, file: FileId, path: PathBuf) {
+        self.file_map.insert(file, path);
+    }
+
+    pub fn add_position(
+        &mut self,
+        pos: PositionId,
+        file: FileId,
+        line: LineNum,
+    ) {
+        self.position_map
+            .insert(pos, SourceLocation::new(line, file));
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct SourceLocation {
+    pub line: LineNum,
+    pub file: FileId,
+}
+
+impl SourceLocation {
+    pub fn new(line: LineNum, file: FileId) -> Self {
+        Self { line, file }
+    }
+}
+
+/// A wrapper around the metadata table that reads file contents into memory.
+/// Constructed via [`MetadataTable::file_reader`].
+///
+///
+/// These allocations are dropped when the reader goes out of scope. Since the
+/// lifetime of the reader is tied to the metadata table that created it, this
+/// isn't intended to be a long term structure.
+pub struct MetadataFileReader<'a> {
+    metadata: &'a MetadataTable,
+    reader_map: RefCell<HashMap<FileId, Box<str>>>,
+}
+
+impl<'a> MetadataFileReader<'a> {
+    pub fn new(metadata: &'a MetadataTable) -> Self {
+        Self {
+            metadata,
+            reader_map: RefCell::new(HashMap::new()),
+        }
+    }
+
+    fn read_file_into_memory(&self, file: FileId) {
+        let path = self.metadata.lookup_file_path(file);
+        let mut reader = std::fs::File::open(path).unwrap();
+        let mut content = String::new();
+        reader.read_to_string(&mut content).unwrap();
+        self.reader_map
+            .borrow_mut()
+            .insert(file, content.into_boxed_str());
+    }
+
+    /// Looks up the given source position. If the file used by this position
+    /// has not been read yet this will cause the contents of the file to be
+    /// read into memory. Will panic if the file does not exist or does not have
+    /// the line number indicated by the position
+    ///
+    /// TODO griffin: make this able to return [str] instead of [String]. Maybe
+    /// also don't buffer file contents into memory? This allocation probably
+    /// isn't a big deal though
+    pub fn lookup_source(&self, pos: &SourceLocation) -> String {
+        let contains_key = self.reader_map.borrow().contains_key(&pos.file);
+        if !contains_key {
+            self.read_file_into_memory(pos.file);
+        }
+
+        let content = &self.reader_map.borrow()[&pos.file];
+
+        let line = content
+            .lines()
+            .nth(pos.line.as_usize())
+            .expect("file does not have the given line number");
+
+        line.to_string()
+    }
+}
