--- calyx-metadata/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "calyx-metadata"
-authors.workspace = true
-license-file.workspace = true
-rust-version.workspace = true
-edition.workspace = true
-version = "0.1.0"
-
-[dependencies]
-cider-idx = { path = "../cider-idx" }
-
-[features]
-serde = ["cider-idx/serde"]
