--- calyx-metadata/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod ids;
-pub mod table;
